+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/RedisPresetLookupDao.java
@@ -12,9 +12,7 @@ import com.dwarfeng.subgrade.stack.exception.DaoException;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.lang.NonNull;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
@@ -77,14 +75,8 @@ public class RedisPresetLookupDao<K extends Key, E extends Entity<K>, JE extends
     }
 
     private List<E> internalEntireLookup(String preset, Object[] objs) {
-        String s = formatter.generalFormat();
-        Set<String> keys = template.keys(s);
-        List<JE> jes = new ArrayList<>();
-        for (String key : keys) {
-            @SuppressWarnings("unchecked")
-            JE je = (JE) template.opsForHash().get(dbKey, key);
-            jes.add(je);
-        }
+        @SuppressWarnings("unchecked")
+        List<JE> jes = template.opsForHash().values(dbKey).stream().map(o -> (JE) o).collect(Collectors.toList());
         List<E> es = jes.stream().map(transformer::reverseTransform).collect(Collectors.toList());
         return es.stream().filter(e -> filter.accept(e, preset, objs)).collect(Collectors.toList());
     }